+++ packages/patternfly-4/react-core/src/components/Checkbox/index.d.ts
@@ -1,1 +1,0 @@
-export { default as Checkbox, CheckboxProps } from './Checkbox';